+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 * #711 Fix click event to switch (thanks @adrlen)
 * #712 Fix SSR support (thanks @anteriovieira)
-* #712 Fix ``header`` slot to tabItem
+* #719 Fix ``header`` slot to tabItem
+* #718 Fix ``changeMonth`` and ``changeYear`` events to datepicker in ``change-month`` and ``change-year``
 
 ## 0.6.4